chinh sua cac truong v1

+++ src/main/java/com/snsapi/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.snsapi.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(int id) {
         super("User not found with ID: " + id);
     }
 }
